updates database

+++ src/database/migrations/2002_01_01_000020_pulsar_create_table_country.php
@@ -12,24 +12,27 @@ class PulsarCreateTableCountry extends Migration
 	 */
 	public function up()
 	{
-            Schema::create('001_002_country', function(Blueprint $table) {
+        if(!Schema::hasTable('001_002_country'))
+        {
+            Schema::create('001_002_country', function (Blueprint $table) {
                 $table->engine = 'InnoDB';
-                $table->string('id_002',2);
-                $table->string('lang_002',2);
-                $table->string('name_002',100);
+                $table->string('id_002', 2);
+                $table->string('lang_002', 2);
+                $table->string('name_002', 100);
                 $table->smallInteger('sorting_002')->unsigned();
-                $table->string('prefix_002',5)->nullable();
-                $table->string('territorial_area_1_002',50)->nullable();
-                $table->string('territorial_area_2_002',50)->nullable();
-                $table->string('territorial_area_3_002',50)->nullable();
+                $table->string('prefix_002', 5)->nullable();
+                $table->string('territorial_area_1_002', 50)->nullable();
+                $table->string('territorial_area_2_002', 50)->nullable();
+                $table->string('territorial_area_3_002', 50)->nullable();
 
-                $table->string('data_lang_002',255)->nullable();
+                $table->string('data_lang_002', 255)->nullable();
                 $table->text('data_002')->nullable();
 
                 $table->primary(['id_002', 'lang_002']);
                 $table->foreign('lang_002', 'fk01_001_002_country')->references('id_001')->on('001_001_lang')
-                        ->onDelete('restrict')->onUpdate('cascade');
+                    ->onDelete('restrict')->onUpdate('cascade');
             });
+        }
 	}
 
 	/**
