+++ src/database/migrations/2002_01_01_000100_pulsar_create_table_resource.php
@@ -12,15 +12,18 @@ class PulsarCreateTableResource extends Migration
 	 */
 	public function up()
 	{
-            Schema::create('001_007_resource', function(Blueprint $table) {
-                $table->engine = 'InnoDB';
-                $table->string('id_007', 30)->primary();
-                $table->string('name_007', 50);
-                $table->integer('package_007')->unsigned();
+		if(!Schema::hasTable('001_007_resource'))
+		{
+			Schema::create('001_007_resource', function (Blueprint $table) {
+				$table->engine = 'InnoDB';
+				$table->string('id_007', 30)->primary();
+				$table->string('name_007', 50);
+				$table->integer('package_007')->unsigned();
 
-                $table->foreign('package_007', 'fk01_001_007_resource')->references('id_012')->on('001_012_package')
-                        ->onDelete('restrict')->onUpdate('cascade');
-            });
+				$table->foreign('package_007', 'fk01_001_007_resource')->references('id_012')->on('001_012_package')
+					->onDelete('restrict')->onUpdate('cascade');
+			});
+		}
 	}
 
 	/**
