+++ src/database/migrations/2002_01_01_000150_pulsar_create_table_preference.php
@@ -12,15 +12,18 @@ class PulsarCreateTablePreference extends Migration
 	 */
 	public function up()
 	{
-		Schema::create('001_018_preference', function(Blueprint $table) {
-			$table->string('id_018', 50)->primary();
-            $table->integer('package_018')->unsigned();
-			$table->text('value_018')->nullable();
-			$table->timestamps();
+		if(!Schema::hasTable('001_018_preference'))
+		{
+			Schema::create('001_018_preference', function (Blueprint $table) {
+				$table->string('id_018', 50)->primary();
+				$table->integer('package_018')->unsigned();
+				$table->text('value_018')->nullable();
+				$table->timestamps();
 
-            $table->foreign('package_018', 'fk01_001_018_preference')->references('id_012')->on('001_012_package')
-                ->onDelete('restrict')->onUpdate('cascade');
-		});
+				$table->foreign('package_018', 'fk01_001_018_preference')->references('id_012')->on('001_012_package')
+					->onDelete('restrict')->onUpdate('cascade');
+			});
+		}
 	}
 
 	/**
