+++ src/database/migrations/2002_01_01_000180_pulsar_create_table_attachment_family.php
@@ -12,19 +12,21 @@ class PulsarCreateTableAttachmentFamily extends Migration {
      */
     public function up()
     {
-        Schema::create('001_015_attachment_family', function(Blueprint $table)
+        if(!Schema::hasTable('001_015_attachment_family'))
         {
-            $table->engine = 'InnoDB';
-            $table->increments('id_015')->unsigned();
-            $table->string('resource_015', 30);                             // resource which belong to this attachment
-            $table->string('name_015', 100);
-            $table->smallInteger('width_015')->unsigned()->nullable();
-            $table->smallInteger('height_015')->unsigned()->nullable();
-            $table->text('data_015')->nullable();
+            Schema::create('001_015_attachment_family', function (Blueprint $table) {
+                $table->engine = 'InnoDB';
+                $table->increments('id_015')->unsigned();
+                $table->string('resource_015', 30);                             // resource which belong to this attachment
+                $table->string('name_015', 100);
+                $table->smallInteger('width_015')->unsigned()->nullable();
+                $table->smallInteger('height_015')->unsigned()->nullable();
+                $table->text('data_015')->nullable();
 
-            $table->foreign('resource_015', 'fk01_001_015_attachment_family')->references('id_007')->on('001_007_resource')
-                ->onDelete('cascade')->onUpdate('cascade');
-        });
+                $table->foreign('resource_015', 'fk01_001_015_attachment_family')->references('id_007')->on('001_007_resource')
+                    ->onDelete('cascade')->onUpdate('cascade');
+            });
+        }
     }
 
     /**
