+++ src/database/migrations/2002_01_01_000200_pulsar_create_table_field_group.php
@@ -12,17 +12,19 @@ class PulsarCreateTableFieldGroup extends Migration {
      */
     public function up()
     {
-        Schema::create('001_025_field_group', function(Blueprint $table)
+        if(!Schema::hasTable('001_025_field_group'))
         {
-            $table->engine = 'InnoDB';
-            $table->increments('id_025')->unsigned();
-            $table->string('name_025', 100)->nullable();
-            $table->string('resource_025', 30);                             // resource which belong to this family field
-            $table->text('data_025')->nullable();
+            Schema::create('001_025_field_group', function (Blueprint $table) {
+                $table->engine = 'InnoDB';
+                $table->increments('id_025')->unsigned();
+                $table->string('name_025', 100)->nullable();
+                $table->string('resource_025', 30);                             // resource which belong to this family field
+                $table->text('data_025')->nullable();
 
-            $table->foreign('resource_025', 'fk01_001_025_field_group')->references('id_007')->on('001_007_resource')
-                ->onDelete('cascade')->onUpdate('cascade');
-        });
+                $table->foreign('resource_025', 'fk01_001_025_field_group')->references('id_007')->on('001_007_resource')
+                    ->onDelete('cascade')->onUpdate('cascade');
+            });
+        }
     }
 
     /**
